Shader Framework
Implemented the movie texture slightly differently, using a material and
a shader. Using premade shaders from the internet changes what the movie
looks like, so it works properly. As of yet, I don't know how exactly to
program the shader, but I'm looking into it. Meanwhile, here is the
framework.