Mooooore Filters!
You can see the list of them in a long comment near the beginning of all
of the filters.
I should probably go to sleep instead of doing all of these.
Oh, they need to be added in the forms of buttons or a list of some sort
or something... But for now you can change the values in startMenu.cs to
see the different ones.